+++ src/components/section/sign-up/email-entry-step.tsx
