+++ src/components/section/sign-up/password-entry-step.tsx
