+++ src/components/section/sign-up/step-indicator.tsx
